Switched from JIT to AOT compilation to prevent occasional first-run slowdowns
